// app/containers/View/types.ts
export interface IViewBase {
  id: number
  name: string
  description: string
  sourceName: string
  [propName: string]: any
}
// ...
export interface IViewLoading {
  view: boolean
  table: boolean
  modal: boolean
  execute: boolean
  copy: boolean
}