// app/containers/Projects/types.ts
export interface IProjectPermission {
  downloadPermission: boolean
  schedulePermission: number
  sharePermission: boolean
  sourcePermission: number
  viewPermission: number
  vizPermission: number
  widgetPermission: number
}

export interface IProject {
  createBy?: { avatar?: string; id?: number; username?: string; email: string }
  permission?: IProjectPermission
  initialOrgId?: number
  userId: number
  isStar?: boolean
  isTransfer?: boolean
  isFavorites?: boolean
  type?: string
  name?: string
  id?: number
  description?: string
  pic?: string
  orgId?: number
  visibility?: boolean
  starNum?: number
  createTime?: string
  updateTime?: string
  isMinePro?: boolean
}